Update registry.json (#196)
Hi, my name is Steven. I'm the SPO for PILOT Pool. I love Cardano and all that it stands for. The community is the best community in all of blockchain in my opinion. This is an exciting time to be in Cardano and this space in general. I truly believe that Cardano will change the world! After much consideration, I am respectfully bowing out of the Single Pool Alliance to hopefully be able to have multiple pools one day (long way off, but with a lot of effort, maybe one day). I'm not sure the best way to do that, so I deleted my original fork, then re-forked the current registry.json file, and deleted my info. Hopefully that does the trick. Is there any more information you need from me? Thank you so much for all that you have done building this alliance, and all that you have done for the community! I support the SPA all the way and I love the cause. I will always be rooting for you! Thank you for all that you have done! The future is exciting! Always, Steven.